tests/test_solver.py
import unittest
from collections import defaultdict
from typing import List

from board.cell import Cell
from board.line import Row, Column
from solver import Solver  # Assuming the `set_all_cells_crossed` method is part of a `Solver` class

class TestSolver(unittest.TestCase):
    def setUp(self):
        """
        Set up the test environment with a mock board and cells.
        """
        # Create mock cells
        self.cell1 = Cell(state="empty")
        self.cell2 = Cell(state="empty")
        self.cell3 = Cell(state="crossed")
        self.cell4 = Cell(state="empty")

        # Create rows and columns
        self.row1 = Row(index=1, cells=[self.cell1, self.cell2, self.cell3])
        self.row2 = Row(index=2, cells=[self.cell4])
        self.col1 = Column(index=1, cells=[self.cell1, self.cell4])
        self.col2 = Column(index=2, cells=[self.cell2])
        self.col3 = Column(index=3, cells=[self.cell3])

        # Assign rows and columns to cells
        for cell in self.row1.cells:
            cell.row_ref = self.row1
        for cell in self.row2.cells:
            cell.row_ref = self.row2
        for cell in self.col1.cells:
            cell.column_ref = self.col1
        for cell in self.col2.cells:
            cell.column_ref = self.col2
        for cell in self.col3.cells:
            cell.column_ref = self.col3

        # Solver instance
        self.solver = Solver()

    def test_set_all_cells_crossed(self):
        """
        Test the `set_all_cells_crossed` method with valid input.
        """
        cells = [self.cell1, self.cell2, self.cell3, self.cell4]

        # Call the method (no implementation yet, so expect no exceptions)
        try:
            self.solver.set_all_cells_crossed(cells)
        except Exception as e:
            self.fail(f"set_all_cells_crossed raised an unexpected exception: {e}")

    def test_invalid_cells(self):
        """
        Test that `set_all_cells_crossed` raises a TypeError for invalid input.
        """
        invalid_cells = [self.cell1, "not a cell", 123]

        with self.assertRaises(TypeError):
            self.solver.set_all_cells_crossed(invalid_cells)

    def test_filter_empty_cells(self):
        """
        Test that only empty cells are considered.
        """
        cells = [self.cell1, self.cell2, self.cell3, self.cell4]
        self.solver.set_all_cells_crossed(cells)

        # Verify that crossed cells are ignored (mock functionality for now)
        filtered_cells = [cell for cell in cells if cell.is_empty()]
        self.assertEqual(len(filtered_cells), 3)

if __name__ == "__main__":
    unittest.main()